prepend path to vended buildkit binaries for Fedora-based systems
