+++ pkg/command/nerdctl_windows.go
@@ -1,0 +1,11 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+//go:build windows
+
+package command
+
+const (
+	envKeyPath       = "Path"
+	envKeyPathJoiner = `\;`
+)
